model-deployment/requirements.txt
streamlit==1.44.0
pandas==2.2.2
seaborn
matplotlib
plotly
pillow==11.2.1
numpy
scikit-learn==1.6.1
